Release 3.0.3.200904291819

+++ org.xmind.cathy/src/org/xmind/cathy/internal/CathyWorkbenchWindowAdvisor.java
@@ -14,22 +14,51 @@
 package org.xmind.cathy.internal;
 
 import java.io.File;
+import java.util.ArrayList;
+import java.util.List;
 
 import org.eclipse.core.runtime.SafeRunner;
 import org.eclipse.jface.action.CoolBarManager;
+import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.jface.util.SafeRunnable;
+import org.eclipse.jface.viewers.ArrayContentProvider;
+import org.eclipse.jface.viewers.LabelProvider;
+import org.eclipse.swt.graphics.Image;
+import org.eclipse.ui.IEditorInput;
+import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.IWorkbenchWindow;
 import org.eclipse.ui.application.ActionBarAdvisor;
 import org.eclipse.ui.application.IActionBarConfigurer;
 import org.eclipse.ui.application.IWorkbenchWindowConfigurer;
 import org.eclipse.ui.application.WorkbenchWindowAdvisor;
+import org.eclipse.ui.dialogs.ListSelectionDialog;
 import org.eclipse.ui.internal.WorkbenchWindow;
 import org.xmind.cathy.internal.actions.SimpleOpenAction;
 import org.xmind.ui.internal.editor.WorkbookEditorInput;
+import org.xmind.ui.internal.editor.WorkbookRefManager;
 import org.xmind.ui.internal.workbench.Util;
+import org.xmind.ui.mindmap.IMindMapImages;
 import org.xmind.ui.mindmap.MindMapUI;
 
 public class CathyWorkbenchWindowAdvisor extends WorkbenchWindowAdvisor {
+
+    private class ListLabelProvider extends LabelProvider {
+        public String getText(Object element) {
+            if (element instanceof IEditorInput)
+                return ((IEditorInput) element).getName();
+            return null;
+        }
+
+        public Image getImage(Object element) {
+            if (element instanceof IEditorInput) {
+                ImageDescriptor image = MindMapUI.getImages().get(
+                        IMindMapImages.XMIND_ICON);
+                if (image != null)
+                    return image.createImage();
+            }
+            return null;
+        }
+    }
 
     public CathyWorkbenchWindowAdvisor(IWorkbenchWindowConfigurer configurer) {
         super(configurer);
@@ -53,7 +82,6 @@ public class CathyWorkbenchWindowAdvisor extends WorkbenchWindowAdvisor {
         super.postWindowOpen();
         final IWorkbenchWindow window = getWindowConfigurer().getWindow();
         if (window != null) {
-
             CoolBarManager coolBar = ((WorkbenchWindow) window)
                     .getCoolBarManager();
             if (coolBar != null) {
@@ -70,39 +98,88 @@ public class CathyWorkbenchWindowAdvisor extends WorkbenchWindowAdvisor {
 
     private void postOpen(final IWorkbenchWindow window) {
         checkLog(window);
+        Object[] lastSession = checkLastSession();
+        if (lastSession != null && lastSession.length > 0) {
+            for (Object input : lastSession) {
+                new WorkbookEditorInput();
+                openEditor((IEditorInput) input, window);
+            }
+        } else {
+            openEditor(null, window);
+        }
+
         window.getShell().getDisplay().asyncExec(new Runnable() {
             public void run() {
-                if (window.getActivePage().getActiveEditor() == null) {
-                    SafeRunner.run(new SafeRunnable() {
-                        public void run() throws Exception {
-                            window.getActivePage().openEditor(
-                                    //new WorkbookEditorInput(),
-                                    new WorkbookEditorInput(),
-                                    MindMapUI.MINDMAP_EDITOR_ID);
-                        }
-                    });
-                }
+                WorkbookRefManager.getInstance().clearLastSession();
             }
         });
+    }
+
+    private void openEditor(final IEditorInput input,
+            final IWorkbenchWindow window) {
+        window.getShell().getDisplay().asyncExec(new Runnable() {
+            public void run() {
+                SafeRunner.run(new SafeRunnable() {
+                    public void run() throws Exception {
+                        IWorkbenchPage activePage = window.getActivePage();
+                        if (input != null) {
+                            activePage.openEditor(input,
+                                    MindMapUI.MINDMAP_EDITOR_ID);
+                        } else {
+                            if (window.getActivePage().getActiveEditor() == null)
+                                activePage.openEditor(
+                                        new WorkbookEditorInput(),
+                                        MindMapUI.MINDMAP_EDITOR_ID);
+                        }
+                    }
+                });
+            }
+        });
+    }
+
+    private Object[] checkLastSession() {
+        List<IEditorInput> session = WorkbookRefManager.getInstance()
+                .loadLastSession();
+        if (session == null || session.isEmpty())
+            return null;
+        ListSelectionDialog dialog = new ListSelectionDialog(null, session,
+                new ArrayContentProvider(), new ListLabelProvider(),
+                WorkbenchMessages.appWindow_ListSelectionDialog_Text);
+        dialog.setTitle(WorkbenchMessages.appWindow_ListSelectionDialog_Title);
+        dialog.setInitialElementSelections(session);
+        dialog.open();
+        return dialog.getResult();
     }
 
     private void checkLog(IWorkbenchWindow window) {
         Log opening = Log.get(Log.OPENING);
         if (opening.exists()) {
-            String[] files = opening.getContents();
-            for (String file : files) {
-                open(window, file);
+            boolean presentation = false;
+            List<String> files = new ArrayList<String>();
+            String[] contents = opening.getContents();
+            for (String line : contents) {
+                if ("-p".equals(line)) { //$NON-NLS-1$
+                    presentation = true;
+                } else
+                    files.add(line);
             }
+            if (files.isEmpty())
+                return;
+            for (String file : files) {
+                open(window, file, presentation);
+                if (presentation)
+                    presentation = false;
+            }
+            files.clear();
             opening.delete();
         }
     }
 
-    private void open(IWorkbenchWindow window, String path) {
+    private void open(IWorkbenchWindow window, String path, boolean presentation) {
         File file = new File(path);
         if (file.isFile() && file.canRead()) {
             window.getShell().getDisplay().asyncExec(
-                    new SimpleOpenAction(window, path));
+                    new SimpleOpenAction(window, path, presentation));
         }
     }
-
 }
